+++ src/components/form/Field/styles.css
@@ -1,0 +1,14 @@
+.Field {
+    display: flex;
+    flex-direction: column;
+    align-items: start;
+    justify-content: start;
+    gap: 0.5rem;
+  }
+  label {
+    font-weight: bold;
+  }
+
+  .Field.inline {
+    flex-direction: row;
+  }
